backend/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Faker==18.9.0
